requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
python-docx==1.1.0
jinja2==3.1.3
python-multipart==0.0.6
aiofiles==23.2.1
